moves the apikeymodal mounting one level down to simplify
